Uploaded Start-Menu-Manager.psh v1.0
<#
.DESCRIPTION
The script operates on a predefined set of registry values under the "HKCU:\Software\Microsoft\Windows\CurrentVersion\Explorer\Advanced" path. Each registry value is intended to control aspects of the Taskbar and Start Menu experience, such as visibility of the Task View button or recommendations in the Start Menu.

A hashtable is used to define the registry values, their expected data types (e.g., DWORD), and the data to be assigned to each. This structure allows for easy extension or modification of the script to accommodate additional registry settings with potentially different types and values.

For each specified registry value, the script performs the following actions:
- Checks if the registry value already exists.
- If the value does not exist, it is created with the specified type and data.
- If the value exists but does not match the expected data, it is updated accordingly.

This approach ensures that the user environment is consistent and aligns with organizational policies, enhancing security and usability.

If any unexpected values are found, the script sends a maintenance alert.
#>